requirements.txt
alembic==1.5.8
blinker==1.4
cffi==1.14.5
click==7.1.2
coverage==5.5
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pyAesCrypt==5.0.0
pycparser==2.20
pycryptodomex==3.10.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pyzipper==0.3.4
six==1.15.0
SQLAlchemy==1.4.2
Werkzeug==1.0.1
WTForms==2.3.3
